fix: first gate run findings -- <cstdint> for INT32_MAX, two sign-compare risks
decode_vorbis.cpp used INT32_MAX without <cstdint> (the only red the docker
host check produced); the variant round-robin compared size_t against uint8_t
and a decode test compared size_t against a signed literal, both of which
-Wextra could turn into -Werror breaks on the cross toolchain.

+++ native/src/audio/audio_engine.cpp
@@ -205,7 +205,7 @@ void AudioEngine::playOn(Bus bus, SoundId id, float gain, float pan,
     }
     std::size_t variant = static_cast<std::size_t>(nextRandom() % count);
     const std::size_t idx = soundIndex(id);
-    if (count > 1 && variant == lastVariant_[idx]) {
+    if (count > 1 && variant == static_cast<std::size_t>(lastVariant_[idx])) {
         variant = (variant + 1) % count;  // never the same twice running
     }
     lastVariant_[idx] = static_cast<std::uint8_t>(variant);
